refactor: store url
