--- app/store/[storeId]/customer/page.js
@@ -1,4 +1,0 @@
-import Customer from "../../../../components/component/dashboard-customer";
-export default function page() {
-  return <Customer />;
-}
