--- app/store/[storeId]/dashboard/page.js
@@ -1,5 +1,0 @@
-import Dashboard from "../../../../components/component/dashboard";
-import Main from "../../../../components/component/dashboard-main";
-export default function dashboard() {
-  return <Main />;
-}
