+++ app/store/customer/page.js
@@ -1,0 +1,4 @@
+import Customer from "../../../components/component/dashboard-customer";
+export default function page() {
+  return <Customer />;
+}
