+++ app/store/inventory/page.js
@@ -1,4 +1,4 @@
-import Inventory from "../../../../components/component/inventory";
+import Inventory from "../../../components/component/inventory";
 
 export default function inventory() {
   return (
