+++ app/store/sale/page.js
@@ -1,9 +1,9 @@
 import React from "react";
-import OrderPage from "../../../../components/order-page-seller/orderPage";
+import Sale from "../../../components/component/sales";
 export default function page() {
   return (
     <div>
-      <OrderPage />
+      <Sale />
     </div>
   );
 }
