+++ app/store/supplier/page.js
@@ -1,4 +1,4 @@
-import Supplier from "../../../../components/component/supplier";
+import Supplier from "../../../components/component/supplier";
 export default function suplier() {
   return (
     <div>
